refactor: move control panel components from /components to /pages/control-panel
