--- src/components/control-panel/model-form/ModelForm.module.css
@@ -1,7 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  margin-bottom: 24px;
-  max-width: 500px;
-}
