--- src/components/control-panel/model-form/index.ts
@@ -1,1 +1,0 @@
-export * from './ModelForm';
