--- src/components/control-panel/model-list/index.ts
@@ -1,1 +1,0 @@
-export * from './ModelList';
