--- src/components/control-panel/request-list/index.ts
@@ -1,1 +1,0 @@
-export * from './RequestList';
