--- src/components/control-panel/user-list/index.ts
@@ -1,1 +1,0 @@
-export * from './UserList';
