+++ src/pages/control-panel/ControlPanel.module.css
@@ -1,3 +1,11 @@
 .addModel {
   margin: 12px 0;
 }
+
+.modelForm {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 24px;
+  max-width: 500px;
+}
